[Comments] CommentActions component
